[CSS]Move img folder to css folder
