+++ StarlingUI-frontend/src/components/deployment-menu/CreateProject.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState, useContext } from 'react';
 import axios from "axios";
 import { ProjectContext } from '../App';
 import { v4 as uuidv4 } from 'uuid';
-import logo from '../img/load.gif';
+import logo from '../../css/img/load.gif';
 const PROJECT_URL = 'http://localhost:8080/design/projects';
 
 export default function CreateProject(props) {
