+++ StarlingUI-frontend/src/css/image.css
@@ -14,7 +14,6 @@
     position: absolute;
     top: .5rem;
     right: 0rem;
-    background: red
 }
 
 .docker-expand{
